feat: made the collapsable list more reusable
