+++ src/components/landing-page/manage-production-button.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router-dom";
 import styled from "@emotion/styled";
-import { SecondaryButton } from "./form-elements";
+import { SecondaryButton } from "../form-elements/form-elements";
 
 const ButtonWrapper = styled.div`
   display: flex;
