+++ src/components/production-list/production-list-components.ts
@@ -1,33 +1,5 @@
 import styled from "@emotion/styled";
-import { isMobile } from "../../bowser";
-import { PrimaryButton, SecondaryButton } from "../landing-page/form-elements";
-import { mediaQueries } from "../generic-components";
-
-export const ProductionItemWrapper = styled.div`
-  text-align: start;
-  color: #ffffff;
-  background-color: transparent;
-  flex: 0 0 calc(25% - 2rem);
-  ${isMobile ? `flex-grow: 1;` : `flex-grow: 0;`}
-  justify-content: start;
-  min-width: 34rem;
-  border: 1px solid #424242;
-  border-radius: 0.5rem;
-  margin: 0 2rem 2rem 0;
-  cursor: pointer;
-
-  ${mediaQueries.isLargeScreen} {
-    flex: 0 0 calc(33.333% - 2rem);
-  }
-
-  ${mediaQueries.isMediumScreen} {
-    flex: 0 0 calc(50% - 2rem);
-  }
-
-  ${mediaQueries.isSmallScreen} {
-    flex: 0 0 calc(100%);
-  }
-`;
+import { PrimaryButton } from "../form-elements/form-elements";
 
 export const ProductionName = styled.div`
   font-size: 1.4rem;
@@ -62,55 +34,6 @@ export const ParticipantCountWrapper = styled.div`
 export const ParticipantCount = styled.div`
   font-size: 1.5rem;
   color: #9e9e9e;
-`;
-
-export const HeaderWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 2rem;
-`;
-
-export const HeaderTexts = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-left: ${({
-    open,
-    isProgramOutputLine,
-  }: {
-    open: boolean;
-    isProgramOutputLine: boolean;
-  }) => (!open && isProgramOutputLine ? "1.5rem" : "0")};
-`;
-
-export const HeaderIcon = styled.div`
-  display: flex;
-  align-items: center;
-  height: 2rem;
-  width: 2rem;
-  flex-shrink: 0;
-`;
-
-export const ProductionLines = styled.div`
-  display: grid;
-  padding: 0 2rem;
-  grid-template-rows: 0fr;
-  transition: grid-template-rows 0.3s ease-out;
-
-  &.expanded {
-    grid-template-rows: 1fr;
-    padding-bottom: 2rem;
-  }
-`;
-
-export const InnerDiv = styled.div`
-  overflow: hidden;
-  display: flex;
-  flex-direction: column;
-  position: relative;
 `;
 
 export const Lineblock = styled.div`
@@ -185,56 +108,15 @@ export const PersonText = styled.div`
   margin-left: 0.5rem;
 `;
 
-export const DeleteButton = styled(SecondaryButton)`
-  display: flex;
-  align-items: center;
-  background: #d15c5c;
-  color: white;
-
-  &:disabled {
-    background: #ab5252;
-  }
-`;
-
 export const CheckboxWrapper = styled.div`
   margin-bottom: 3rem;
   margin-top: 0.5rem;
 `;
 
-export const ManageButtons = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  margin: 1rem 0 1rem 0;
-`;
-
-export const NameEditButton = styled.button`
-  background: transparent;
-  border: none;
-  padding: 0;
-  cursor: pointer;
-  font: inherit;
-  display: flex;
-  justify-content: center;
-  align-self: center;
-  margin: 1rem 0;
-  flex-shrink: 0;
-  height: 2rem;
-  width: 2rem;
-
-  svg {
-    width: 100%;
-    height: 100%;
-  }
-
-  &:hover svg {
-    transform: scale(1.2);
-  }
-`;
-
 export const AddLineSectionForm = styled.form`
   margin-top: 1rem;
   border: 1px grey solid;
-  border-radius: 5px;
+  border-radius: 0.5rem;
   padding: 1rem;
   position: relative;
 `;
@@ -260,12 +142,6 @@ export const RemoveIconWrapper = styled.div`
   }
 `;
 
-export const SpinnerWrapper = styled.div`
-  position: relative;
-  width: 2rem;
-  height: 2rem;
-`;
-
 export const IconWrapper = styled.div`
   height: 2rem;
   width: 2rem;
@@ -277,18 +153,6 @@ export const IconWrapper = styled.div`
   svg {
     height: 100%;
     width: 100%;
-  }
-`;
-
-export const EditNameWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: start;
-  max-width: 30rem;
-  pointer-events: none;
-
-  > * {
-    pointer-events: auto;
   }
 `;
