+++ src/components/share-line-link/share-line-link-modal.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useRef, useState } from "react";
-import { DecorativeLabel, FormInput } from "../landing-page/form-elements";
+import { DecorativeLabel, FormInput } from "../form-elements/form-elements";
 import { Modal } from "../modal/modal";
 import { CopyButton } from "../copy-button/copy-button";
 import { RefreshButton } from "../refresh-button/refresh-button";
@@ -81,7 +81,10 @@ export const ShareLineLinkModal = ({
                   {isCopyProduction && (
                     <DecorativeLabel>{name}</DecorativeLabel>
                   )}
-                  <FormInput value={isLoading ? "Loading..." : urlToDisplay} />
+                  <FormInput
+                    readOnly
+                    value={isLoading ? "Loading..." : urlToDisplay}
+                  />
                 </LinkLabel>
                 <CopyButton
                   urls={[urlToDisplay]}
