+++ src/components/shared/use-handle-header-click.ts
@@ -1,0 +1,17 @@
+import { useCallback } from "react";
+
+export const useHandleHeaderClick = (editNameOpen: boolean) => {
+  const handleHeaderClick = useCallback(
+    (e: React.MouseEvent, open: boolean, setOpen: (open: boolean) => void) => {
+      if (
+        !editNameOpen &&
+        !(e.target as HTMLElement).closest(".name-edit-button")
+      ) {
+        setOpen(!open);
+      }
+    },
+    [editNameOpen]
+  );
+
+  return handleHeaderClick;
+};
